Revise error handling in node mgr again
Finally this makes the most sense. It is *not* a user reportable error,
but it *should error* not panic! The error is recoverable in the sense
that we can ignore it and write to the error log. Down the line it might
be a good idea to also shoot something generic to the UI such that the
user can report the bug.