Add Ramp and Grayscale nodes
Grayscale works with one single mode now, there are other weightings and
conversion methods. There's a post about this somewhere on the internet
but I can't find it at the moment.

The Ramp doesn't do anything at all for now.

Both of these nodes are required for manual type coercion if nothing else.