Remove Rgba ImageType
Any processing is better done on RGB than on RGBA. If an alpha channel
is needed in a PBR pipeline, it must be provided separately to the
shader anyhow. It does not make sense to have it as part of any PBR map.
For flexibility the *output* stage can still support arbitrary swizzling
from multiple images. In fact, it should. See Mixer or Designer for
prior art.