OutputReady event
When an output is ready, the compute manager emits this event. It is
primarily intended to be caught by the render manager, which will then
forward it to the renderer in question. To do this, the render manager
will have to know which renderer is responsible for which set of images.

The BrokerImageView struct is a rather unsafe construct. As far as
gfx-hal is concerned, it is perfectly safe to send ImageViews between
threads. However, in order to do so we hide the backend parameter. Not
doing so would mean incurring the parameter on the entire Lang struct,
which in turn messes with the UI TLS. Since there is only ever one
backend in the entire program, this is fine in practice.