Fix cache invalidation for multiple successors
The old algorithm was inherently incapable of dealing with this. Instead
of using a bool, we now record the "generation" of each image in a u64,
and compare that. This is technically bounded, but a u64 is large enough
that if each full graph computation (e.g. on parameter change) takes
1ms, it'd take about 1.4x the age of the oldest known land animal until
we run out of numbers. I think that'll do.