Limit min memory size requests to chunk size
We need to allocate at least one chunk in any case. For very small
images (<= 128x128 rgba16f with the current chunk size), this may be
mildly wasteful, and cause fragmentation. But neither of these should
turn out to be a big problem. The remedy here would be to lower the
chunk size.