Rework OperatorParamBox trait
This decouples the UI from the operator definition even more. In order
to do so, we need to use heap allocated collections in the descriptions
and cannot use plain slices anymore. This may lead to extra allocs, but
those shouldn't be a problem in the real world. Furthermore, we need
dynamic allocation for complex operators either way.