Adapt renderer to render to in-memory image
Since all drawing and swapchain handling is now done by the UI and we no
longer draw to a compositor surface directly, the renderer now contains
a target image to which it renders. Probably.

This image(view) must be transferred to the UI thread, where it will
then be blitted on during UI rendering in the right spot.

Resizing logic outstanding.