Mutate parameters on widget change
This is required for the widgets to reflect change. Note that this does
mutate the parameter box description, *not* the operator! This is easier
and faster because it doesn't require any roundabout way to read and
write to the operator through the parameter description. This means that
it's now likely superfluous to actually keep the whole operator around
in the graph