Draft thumbnail storage
This is memory managed by the compute component into which thumbnails
are placed. Previously we sent the whole thumbnail data over the broker
channel to the UI. With the new UI, we have direct access to GPU data
during rendering and therefore only need to transport an ImageView
handle. To that end, we now keep all thumbnail data on the GPU at all
times. This data is relatively small (512 thumbnails taking ~32MB). The
cache structure in this commit is built to expand on demand. It does not
do any referencing between what thumbnail image matches which resource!