Hack fix descriptor sets in UI
This is a broken design and needs to be fixed. Ideally the descriptors
should just be kept with the image map itself, and the renderer offers
up methods to mutate the image map. Otherwise, synchronization between
that and the descriptor map is a nightmare, and it's all broken, and
because of the ugly hack that BrokerImageView is, apparently we even get
corrupted data on rebuild.