Fix crash involving reuse of old nodes
This is a tricky one, because it shows that widget state may be retained
arbitrarily in conrod. I don't see any way to manually discard a widget
either, and in fact you manual care would have to be taken to also
discard the old Id anyhow. We'll see how this plays out, in principle
this looks like a memory leak waiting to happen.