Implement reuse of socket ids between UI nodes
This is more consistent with conrod's overall design philosophy, and
necessary for switching between graphs. The implementation should be
relatively efficient, relying on hashing of the sockets on build, and
caching the "correct" hash. Rebuilding *should* preserve socket ids on
grow. It does not preserve them on shrink, so in a sense we are still
leaking widget ids here.