Handle inputs in call
This doesn't work quite the way I wanted it to. Because input nodes have
output sockets (duh), they automatically receive *TypedOutputs* in the
compute manager. This means that they have an actual GPU Image
associated with them, and in turn this means we can't just "rename" an
existing output image via move, like we can do with input sockets.

Therefore this performs a copy. In effect this means we do call by value
rather than call by reference, if you will. This is a performance issue,
although likely a minor one in the grand scheme of things. Blitting is
very fast after all.