Fix segfault on load of subgraphs with inputs
The reason for this was that the target image of the copy operation (on
call) was unallocated, and therefore Vk was happily doing undefined
things. The downside of this fix is that now we check for alloc every
time we copy. However, this is cheap in the grand scheme of things.