Relinearize graph on switching away
This makes sure that there's always an up to date copy, e.g. after
changing a name and switching away. Before this would crash the compute
component, as the rename would work there but not be part of the
linearization, leading to crashes while working through the stored
linearization on call.