Add TopoSort linearization mode
This happened in the process of determining the last usage point for any
given node during linearization. A topological sort is sufficient as
long as there's enough VRAM that we never have to evict prior results
that are still needed. This should *generally* be the case unless
working on *huge* graphs or with very little VRAM.

The full traversal is still part of the codebase and can easily be
switched to by changing the parameter. In the future it may make sense
to either switch on demand or expose it as a user setting for low end
machines (or particularly pathological graphs). It should be possible
however to trigger a full traversal automatically if image cleanup fails
in the compute component.