Introduce Use action in linearization
The idea here is that in order to find the last use, we denote use
points in the stack explicitly. Previously we relied on noting use
points during visit, but this does rather mark "creation" points instead
of use points. The two traversals over the edges are required to make
sure that all use points happen at the same step before visiting the
user node.

This *should* be a correct algorithm now, but no proof has been
performed. It looks correct on examples so far.