Reuse reinit_image for cleanup
It turns out that the Vulkan spec prohibits images from being rebound.
"Non-sparse resources are backed by a single physical allocation prior
to device use (via vkBindImageMemory or vkBindBufferMemory), and their
backing must not be changed." Therefore we have to recreate the image
itself anyway, and thus we just drop the entire thing and rebuild it. It
will of course only alloc when required.