Use Arc Mutex for transferable image views
This uses dynamic typing on the Lang side and doesn't incur a <B>
parameter. This patch fixes the problem of render views not showing up
in release builds, as it gets rid of the hack used to transfer images
before, which seemed to have been undefined behaviour.

That said, this isn't perfect yet. try_unwrap should really be repeated
in case we can't acquire the data. Right now this is a race condition!