Remove FIXME
The race condition seems fixed, and should be fixed, since the render
manager holds a lock throughout the entire copying process. However, a
bug that *looks* the same is now occurring with thumbnail generation
somehow. Not sure whether that one is a race condition or not, or
whether it's new or was just masked out before.