Fix freeing of Arc'd GPU resources
The previous method was executing ManuallyDrop::take multiple times when
there was contention. This would corrupt the heap, which in turn would
crash malloc somewhere in Vulkan implementation code.

We also have to make sure that we only lock the GPU once the other
resources have been acquired. Failing to do so can cause a deadlock as
another thread still tries to acquire the GPU to do its work!