Basic refactoring for layer/graph support in UI
This mirrors some of the backend changes, introducing a trait for common
actions. Since most of the graph specific stuff is already hidden behind
finding the target graph first, not much has to change.

node_graph in App will panic now when called unless a graph is actually
active. This should be fine.