Rework cleanup algorithm
This fixes a case in which memory was cleaned up from within a
call to a subgraph despite being required later. This method also cleans
up anything older than the current execution. It is therefore overeager
and if required repeatedly, this is far from ideal because it throws off
all the caching.

It is important to eventually implement a scheme that at least *prefers*
to keep expensive images around.