Fix cache invalidation on calls
This is a bug that occurred when calling subgraphs multiple times where
images wouldn't be processed because they were already processed in a
prior call and the seq number didn't reflect that. This bug came in
during a refactor, and this restores the original behaviour.
This *should* work fine, since seq numbers can grow arbitrarily large,
and the caching scheme only depends on comparing op_seq and input seqs.