Improve normal estimation
Part of this is using the data directly from the heightfield, with a
more appropriate (not hard coded) sample step dependent on view distance
and texel size. But an even bigger part is using 32 bit height data in
the renderer. This commit uses 32 bit floating point images for *all*
channels, which is inappropriate.