Improve AO method
This is now utilizing the temporal multisampling implemented for AA. The
method raymarches a cone out from the surface and checks at each step
how much of the cone is intersected by the geometry. Of course because
the SDF is not a proper SDF, this is approximate. Through multisampling
we get what *looks* to be a decent AO estimate however.

Since this is reasonably expensive (4 cones per sample), we should still
strive to provide a secondary cheap (i.e. fit for real-time)
appromixation, such as SSAO.