Add BasicImageBuilder and use in environment
This builder can be used to build basic images with memory and view.
Basic in the sense that each image is backed by its own adequately sized
memory. This is used in a bunch of places such as environment maps. The
most notable exception is of course the compute memory region.