Remove TRANSIENT flag from UI cmd buffer
Since this buffer is kept around and reset per frame, this should be
better. The alternative is allocating a fresh buffer every time, which
might be more expensive as memory is taken from and returned to the
system every frame.

Might make sense to think about buffers like this for other places too.
Renderer comes to mind