Fix output removal bug
The problem seems to be one of missing reindexing. The =NodeGraph= keeps an index map around, which maps node names (resources essentially) to node indices in the underlying graph. When deleting a node, this index structure has to be updated because the indices in the graph change. It also keeps an outputs list, /which currently does not get updated/. Hence once Perlin Noise is deleted, the outputs list has a node index that doesn't exist anymore. Subsequent resource lookup will then fail on =.unwrap()=.