Add extra options to Distance
Not quite final perhaps, extent might be hard to use since it's
exponential. Linear distance amounts however are hard to control in the
lower ranges and that's where it intuitively seems to matter the most.
Perhaps having controls that allow changing from exp to linear scale etc
would be the best solution here.