Implement File IO in IO component
The IO component is to take care of all surface file related IO. It
intercepts the user events to do IO, manages internal state related to
it, and then dispatches back into the system.

On saving, *other* components only interpret their part of the
instruction and publish their data. The IO component takes this data
from the bus and integrates it into the surface file to be assembled.
When the file is complete, it stores it to disk