Fix thumbnail seq conditions
This was a problem introduced by skipping thumbnail generation in called
graphs. When inspecting the called graphs, even a recompute wouldn't
force thumbnails, because thumbnails are only computed for images
updated in the current seq.

Switching this to current or older would generate thumbnails for old
images that haven't been affected by any change.

Therefore we keep track of the thumbnail seq separately and update
whenever the thumbnail seq does not match the image seq.