Terminate sphere trace on exit of bounding volume
This fixes a problem with matcaps, but also an issue that occurred at
times with the PBR renderer when the normals were broken. The additional
branch might be costly, we'll have to check that once we get around to
optimizing. In principle though this should allow early termination of
rays which is almost always preferable.