Disable most of graph widget
Preparing for (partial) rewrite on R-Tree basis. It seems like we don't
need a graph at all in the frontend, instead what we really need are
spatial relations, since that's where most of our queries are likely to
go too. Having an extra R-Tree as an acceleration structure would work,
but there's a lot of duplication going on in that case.