Add Alpha Extract node
Usually we don't have an alpha channel and simply ignore it. It only
exists because of Vulkan format requirements not requiring rgb16f, so we
sacrifice those bytes. When uploading images or using SVGs however, the
Alpha channel may become relevant. For this we now have the alpha
extract node, to convert it into its own grayscale map for further processing.