Add basic scaffolding for occupancy buffer
I.e. a buffer passed to the shader describing which inputs are used and
what ImageType is on each input. This should be optional for each
shader. Such a struct will allow optional inputs on atomics (complex
operators have to be handled separately, through default values coming
out of the input node) as well as shaders deciding how to deal with data
depending on input types (e.g. Select node)